refactor: move city slicer to weather folder
